fixed compilation by using correct type: strlen() returns size_t, not int
Side note: why do we require that function newVariable() is called
with a string and its length (implicitly requiring that the correct
length is provided?), while the string length could easily be determined
inside the function ???

+++ monetdb5/mal/mal_instruction.c
@@ -751,7 +751,7 @@ makeVarSpace(MalBlkPtr mb)
 
 /* create and initialize a variable record*/
 int
-newVariable(MalBlkPtr mb, str name, int len, malType type)
+newVariable(MalBlkPtr mb, str name, size_t len, malType type)
 {
 	int n;
